be more compatible with webdav: on rename, modify both displayName and path id
on this subject:

http://community.nuxeo.com/5.3/books/nuxeo-book/html/nuxeo-platform-webdav-server.html

39.1.1. Path vs displayName

In WebDAV specification each accessible resource must have a name (part of the URL) and can have a displayName
(defines how the resource should be displayed to the user).
Unfortunately, most WebDAV client don't use the displayName property even if they ask the server to send it.
Because of this limitations, most client use the last part of the URL as display name (with all limitations due to URL encoding).
In addition of the display problem, this has side effect on other features like renaming.
When trying to rename a resource (for example after creating a new Collection/Folder),
most clients will send a move operation (change the path) instead of sending a propatch on displayName.